+++ Sources/SwiftCurrent_SwiftUI/Protocols/WorkflowModifier.swift
@@ -1,0 +1,15 @@
+//
+//  WorkflowModifier.swift
+//  SwiftCurrent_SwiftUI
+//
+//  Created by Tyler Thompson on 8/21/21.
+//  Copyright © 2021 WWT and Tyler Thompson. All rights reserved.
+//
+
+import SwiftUI
+import SwiftCurrent
+
+@available(iOS 14.0, macOS 11, tvOS 14.0, watchOS 7.0, *)
+protocol WorkflowModifier {
+    func modify(workflow: AnyWorkflow)
+}
